integrasi category fe to be done

+++ app/Http/Controllers/CategoryController.php
@@ -9,37 +9,35 @@ use Illuminate\Support\Facades\File;
 class CategoryController extends Controller
 {
     function index(){
-        $data = Category::all();
-        $jumlah_produk=0;
-        return view("category.category", compact("data", "jumlah_produk"));
-    }   
-
-    function create(){
-        return view("category.create");
-    }
-
+        $categories = Category::withCount("product")->get();
+        return view("admin.manajemen-kategori", compact("categories"));
+    }  
+    
     function store(Request $request){
         $data =$request->validate([
             "name"=>["required"],
             "description"=>["required"],
             "icon"=>["required","image", "mimes:jpeg,png,jpg", "max:2048"],
+            "banner" => ["nullable", "image", "mimes:jpeg,png,jpg", "max:2048"],
         ]);
 
         $extension = $request->icon->extension();
         $imgname = date('dmyHis') . '.' . $extension;
         $image = $request->file("icon");
-        $image->storeAs('public/categories', $imgname, "public");
+        $path = $image->storeAs('categories/icon', $imgname, "public");
+        $data["icon"] = "storage/" . $path;
 
-        $data["icon"] = 'categories/'.$imgname;
+        if(isset($data["banner"])){
+            $extension = $request->icon->extension();
+            $imgname = date('dmyHis') . '.' . $extension;
+            $image = $request->file("banner");
+            $path = $image->storeAs('categories/banner', $imgname, "public");
+            $data["banner"] = "storage/" . $path;
+        }
     
         Category::create($data);
         
-        return redirect()->route("category.index");
-    }
-
-    function edit($id){
-        $category = Category::find($id);
-        return view("category.edit", compact("category"));
+        return redirect()->route("manajemen-kategori")->with("add-success", "category berhasil ditambah");
     }
 
     function update(int $id,Request $request){
@@ -47,13 +45,12 @@ class CategoryController extends Controller
             "name" => ["required"],
             "description" => ["required"],
             "icon" => ["nullable", "image", "mimes:jpeg,png,jpg", "max:2048"],
+            "banner" => ["nullable", "image", "mimes:jpeg,png,jpg", "max:2048"],
         ]);
-
         $category = Category::find($id);
 
         if(isset($data["icon"])){
-
-            if ($category->sampul_buku) {
+            if ($category->icon) {
                 try {
                     if (File::exists("/" . $category->icon)) {
                         File::delete("/" . $category->icon);
@@ -66,21 +63,42 @@ class CategoryController extends Controller
             $extension = $request->icon->extension();
             $imgname = date('dmyHis') . '.' . $extension;
             $image = $request->file("icon");
-            $path= $image->storeAs('/categories', $imgname, "public");
-            $data["icon"]=$path;
+            $path= $image->storeAs('categories/icon', $imgname, "public");
+            $data["icon"]= "storage/".$path;
 
+
+    
+        }
+        if (isset($data["banner"])) {
+            if ($category->banner) {
+                try {
+                    if (File::exists("/" . $category->banner)) {
+                        File::delete("/" . $category->banner);
+                    }
+                } catch (\Throwable $th) {
+                    return back()->withError("gagal menyimpan");
+                }
+            }
+
+            $extension = $request->icon->extension();
+            $imgname = date('dmyHis') . '.' . $extension;
+            $image = $request->file("banner");
+            $path = $image->storeAs('categories/banner', $imgname, "public");
+            $data["banner"] = "storage/" . $path;
         }
 
         $category->update($data);
         $category->save();
 
-        return redirect()->route("category.index")->with("success", "category berhasil diganti");
+        return redirect()->route("manajemen-kategori")->with("update-success", "category berhasil diganti");
     }
 
     function delete($id){
-        $category = Category::find($id);
+        $category = Category::where("id", $id)->withCount("product")->first();
+        if($category->product_count < 0){
+            return redirect()->route("manajemen-kategori")->with("error", "category gagal dihapus");
+        }
         $category->delete();
-
-        return redirect()->route("category.index")->with("success", "category berhasil dihapus");
+        return redirect()->route("manajemen-kategori")->with("delete-success", "category berhasil dihapus");
     }
 }
